Rewrite core initial guess
